Move all the jeedrunner files into a jeedrunner package
