+++ demo/src/main/kotlin/edu/illinois/cs/cs125/answerablewebdemo/Main.kt
@@ -1,6 +1,9 @@
 package edu.illinois.cs.cs125.answerablewebdemo
 
 import edu.illinois.cs.cs125.answerable.*
+import edu.illinois.cs.cs125.answerable.jeedrunner.answerableBytecodeProvider
+import edu.illinois.cs.cs125.answerable.jeedrunner.jeedOutputCapturer
+import edu.illinois.cs.cs125.answerable.jeedrunner.jeedSandbox
 import edu.illinois.cs.cs125.jeed.core.*
 import io.ktor.application.*
 import io.ktor.features.ContentNegotiation
@@ -12,6 +15,8 @@ import io.ktor.routing.*
 import io.ktor.server.engine.*
 import io.ktor.server.netty.*
 import kotlin.random.Random
+
+// Essentially this entire module is a TODO
 
 fun main1() {
 
@@ -48,6 +53,7 @@ data class AnswerableDemoPost(
     val solutionName: String
 )
 
+// TODO: Take advantage of the jeedrunner service
 private fun main() {
     val commonSource = Source(mapOf(
             "Common.java" to """
