+++ jeedrunner/src/main/kotlin/edu/illinois/cs/cs125/answerable/jeedrunner/Adapter.kt
@@ -1,4 +1,4 @@
-package edu.illinois.cs.cs125.answerable
+package edu.illinois.cs.cs125.answerable.jeedrunner
 
 import edu.illinois.cs.cs125.answerable.api.BytecodeClassProvider
 import edu.illinois.cs.cs125.answerable.api.BytecodeProvider
