+++ jeedrunner/src/main/kotlin/edu/illinois/cs/cs125/answerable/jeedrunner/Main.kt
@@ -1,4 +1,4 @@
-package edu.illinois.cs.cs125.answerable
+package edu.illinois.cs.cs125.answerable.jeedrunner
 
 import java.io.File
