+++ jeedrunner/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-edu.illinois.cs.cs125.answerable.AnswerableAnnotationOwner
+edu.illinois.cs.cs125.answerable.jeedrunner.AnswerableAnnotationOwner
