+++ jeedrunner/src/test/kotlin/edu/illinois/cs/cs125/answerable/jeedrunner/TestCorrectness.kt
@@ -1,5 +1,6 @@
-package edu.illinois.cs.cs125.answerable
+package edu.illinois.cs.cs125.answerable.jeedrunner
 
+import edu.illinois.cs.cs125.answerable.ExecutedTestStep
 import edu.illinois.cs.cs125.answerable.classdesignanalysis.Matched
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
