+++ jeedrunner/src/test/kotlin/edu/illinois/cs/cs125/answerable/jeedrunner/TestSandbox.kt
@@ -1,5 +1,6 @@
-package edu.illinois.cs.cs125.answerable
+package edu.illinois.cs.cs125.answerable.jeedrunner
 
+import edu.illinois.cs.cs125.answerable.ExecutedTestStep
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
