+++ jeedrunner/src/test/kotlin/edu/illinois/cs/cs125/answerable/jeedrunner/TestService.kt
@@ -1,7 +1,6 @@
-package edu.illinois.cs.cs125.answerable
+package edu.illinois.cs.cs125.answerable.jeedrunner
 
-import edu.illinois.cs.cs125.answerable.jeedrunner.Answerable
-import edu.illinois.cs.cs125.answerable.jeedrunner.QuestionLanguage
+import edu.illinois.cs.cs125.answerable.AnswerableMisuseException
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
